Update for Home Remote version 2.12.0
This update is to match the new version of Home Remote - 2.12.0.  

Home Remote version 2.12.0's main upgrade is that "plugin.Devices" is now a JavaScript object instead of a JavaScript array.   This was changed because it was discovered that looping through large arrays can take a while & have a very significant impact on performance.  

https://groups.google.com/forum/#!msg/thehomeremote/bjtILNK6mJ0/YtHVB1ArHAAJ 

Personally,  I have found a significant decrease in the time the app takes to parse the data received from Hubitat.  

This code also adds a new hard coded device - "Connection".   The state of the device can be displayed in the app to show whether you are connected locally or via the cloud.  It has other uses as well - I use it to decide which web address to use to access my webcams.  

Also note I have changed the polling interval:  plugin.PollingInterval = 500;
Home Remote will now request changes from Hubitat every 500ms.  This may be too often for some users.  I'm also not sure if it will cause Hubitat to get bogged down.  Let me know!